Fix multi-shard SELECT problem with HAVING clauses
Because Vars from HAVING clauses weren't being included in the remote
query's target list, they would be NULL if they did not appear in the
target list or a GROUP/ORDER clause. This change includes them in the
remote target list so they'll be available for local evaluation of the
HAVING clause.