Detect zero-shard queries and use standard plan
It's possible certain queries involve zero shards. For example, SELECTs
with conflicting conditions (such as col = n AND col =m) in their WHERE
clauses will involve zero shards. In this case, we swap back in the
original plan and just let the query hit the local table on the master.
This produces correct results while avoiding the hit of remote queries
or temp tables.